moved tests into separate pkg.
